Breaking up Engine/Snapshots
