+++ src/Engine/Graphics/DecorationList.cpp
@@ -3,8 +3,6 @@
 #include <cstdlib>
 
 #include "Engine/Engine.h"
-#include "Engine/Snapshots/EntitySnapshots.h"
-#include "Engine/Snapshots/SnapshotSerialization.h"
 #include "Level/Decoration.h"
 
 #include "Utility/String.h"
