+++ src/Engine/Graphics/ImageLoader.cpp
@@ -9,11 +9,11 @@
 #include "Engine/Graphics/IRender.h"
 #include "Engine/Graphics/Sprites.h"
 #include "Engine/Graphics/Texture_MM7.h"
-#include "Engine/Snapshots/CommonSnapshots.h"
 #include "Engine/LodTextureCache.h"
 #include "Engine/LodSpriteCache.h"
 #include "Engine/LOD.h"
 
+#include "Library/Snapshots/CommonSnapshots.h"
 #include "Library/Image/ImageFunctions.h"
 #include "Library/Image/PCX.h"
 #include "Library/Logger/Logger.h"
