+++ src/Engine/Graphics/Sprites.cpp
@@ -7,10 +7,6 @@
 #include "Engine/Engine.h"
 #include "Engine/EngineIocContainer.h"
 #include "Engine/OurMath.h"
-
-#include "Engine/Snapshots/SnapshotSerialization.h"
-#include "Engine/Snapshots/CompositeSnapshots.h"
-
 #include "Engine/Graphics/DecorationList.h"
 #include "Engine/Graphics/PaletteManager.h"
 #include "Engine/Graphics/Image.h"
