+++ src/Engine/Graphics/TextureFrameTable.cpp
@@ -1,7 +1,5 @@
 #include "TextureFrameTable.h"
 
-#include "Engine/Snapshots/EntitySnapshots.h"
-#include "Engine/Snapshots/SnapshotSerialization.h"
 #include "Engine/AssetsManager.h"
 
 #include "Utility/String.h"
