+++ src/Engine/Snapshots/CMakeLists.txt
@@ -6,12 +6,10 @@ set(ENGINE_SERIALIZATION_SOURCES
         TableSerialization.cpp)
 
 set(ENGINE_SERIALIZATION_HEADERS
-        CommonSnapshots.h
         CompositeSnapshots.h
         EntitySnapshots.h
-        SnapshotSerialization.h
         TableSerialization.h)
 
 add_library(engine_serialization STATIC ${ENGINE_SERIALIZATION_SOURCES} ${ENGINE_SERIALIZATION_HEADERS})
-target_link_libraries(engine_serialization PUBLIC engine library_binary)
+target_link_libraries(engine_serialization PUBLIC engine library_binary library_snapshots)
 target_check_style(engine_serialization)
