+++ src/Engine/Snapshots/CompositeSnapshots.cpp
@@ -17,7 +17,7 @@
 #include "Engine/LOD.h"
 #include "Engine/Party.h"
 
-#include "CommonSnapshots.h"
+#include "Library/Snapshots/CommonSnapshots.h"
 
 void reconstruct(const IndoorLocation_MM7 &src, IndoorLocation *dst) {
     reconstruct(src.vertices, &dst->pVertices);
