+++ src/Library/CMakeLists.txt
@@ -11,4 +11,5 @@ add_subdirectory(Lod)
 add_subdirectory(Logger)
 add_subdirectory(Random)
 add_subdirectory(Serialization)
+add_subdirectory(Snapshots)
 add_subdirectory(Trace)
